+++ packages/nextclaw-core/src/index.ts
@@ -10,6 +10,7 @@ export * from "./bus/queue.js";
 export * from "./channels/feishu-probe.js";
 export * from "./channels/manager.js";
 export * from "./channels/plugins/index.js";
+export * from "./commands/registry.js";
 export * from "./config/brand.js";
 export * from "./config/actions.js";
 export * from "./config/loader.js";